--- gobook-shop-service/src/main/java/me/jincrates/shopservice/config/WebConfig.java
@@ -1,4 +1,0 @@
-package me.jincrates.shopservice.config;
-
-public class WebConfig {
-}
