--- gobook-shop-service/src/main/java/me/jincrates/shopservice/controller/item/ItemController.java
@@ -1,4 +1,0 @@
-package me.jincrates.shopservice.controller.item;
-
-public class ItemController {
-}
