--- gobook-shop-service/src/main/java/me/jincrates/shopservice/domain/enumtype/ItemSellStatus.java
@@ -1,5 +1,0 @@
-package me.jincrates.shopservice.domain.enumtype;
-
-public enum ItemSellStatus {
-    SELL, SOLD_OUT
-}
