--- gobook-shop-service/src/main/java/me/jincrates/shopservice/domain/enumtype/MemberRole.java
@@ -1,5 +1,0 @@
-package me.jincrates.shopservice.domain.enumtype;
-
-public enum MemberRole {
-    USER, ADMIN
-}
