+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/infra/common/entity/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package me.jincrates.shopservice.domain;
+package me.jincrates.shopservice.infra.common.entity;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
