+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/infra/config/WebConfig.java
@@ -1,0 +1,4 @@
+package me.jincrates.shopservice.infra.config;
+
+public class WebConfig {
+}
