+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/infra/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package me.jincrates.shopservice.exception;
+package me.jincrates.shopservice.infra.exception;
 
 public record ErrorResponse(Integer code, String message) {
