+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/infra/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package me.jincrates.shopservice.exception;
+package me.jincrates.shopservice.infra.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ExceptionHandler;
