+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/module/item/api/ItemApi.java
@@ -1,0 +1,4 @@
+package me.jincrates.shopservice.module.item.api;
+
+public class ItemApi {
+}
