+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/module/item/domain/ItemSellStatus.java
@@ -1,0 +1,5 @@
+package me.jincrates.shopservice.module.item.domain;
+
+public enum ItemSellStatus {
+    SELL, SOLD_OUT
+}
