+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/module/item/web/dto/ItemDto.java
@@ -1,0 +1,4 @@
+package me.jincrates.shopservice.module.item.web.dto;
+
+public class ItemDto {
+}
