+++ gobook-shop-service/src/main/java/me/jincrates/shopservice/module/member/domain/MemberRole.java
@@ -1,0 +1,5 @@
+package me.jincrates.shopservice.module.member.domain;
+
+public enum MemberRole {
+    USER, ADMIN
+}
