--- gobook-shop-service/src/main/java/me/jincrates/shopservice/web/dto/ItemDto.java
@@ -1,4 +1,0 @@
-package me.jincrates.shopservice.web.dto;
-
-public class ItemDto {
-}
